recipes-kernel/linux: allow the usage of the cip-kernel-config
To avoid the duplication of the linux cip kernel configuration
add the option USE_CIP_KERNEL_CONFIG. If this option is set to 1
the KERNEL_DEFCONFIG from the cip project is used. The KERNEL_DEFCONFIG
is contains the explicit path to the defconfig in
gitlab.com/cip-project/cip-kernel-config.git.

SRCREV_cip-kernel-config sets the commit sha for checkout.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
[Jan: reformatted style of linux-cip-common.inc]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>