Uprevision the ISAR to latest one
Update to latest ISAR.

When we build the images with efibootguard wic plugin it is expecting
the KERNEL_IMAGE and INITRD_IMAGE variable, but it is removed in
Upstream ISAR due to cleanup the unused variables.

Ref: https://github.com/ilbers/isar/commit/6fa0197ac36a86b2ae5c27f5b7804dd9a46d3c0b

But those variables we used in downstream layer with efibootguard wic
plugin hence added required WIC variables in efibootguard.yml to genereate successful wic images.

Signed-off-by: Srinuvasan A <srinuvasan_a@mentor.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>