deploy-cip-core.sh: Start organising cip-core uploads by branch
At the moment we only deploy binaries if building for the master branch.
For development and testing it would be useful to have access to the
binaries built for all branches.

This patch starts storing the binaries by branch/tag name
(CI_COMMIT_REF_SLUG).

This allows users to be able to download the latest master, next or
development binaries as required.

Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>

---

We may want to consider setting an expiry lifecycle rule in the s3 bucket
so that we don't end up with a lot of old files for old development
branches. Master and next branches should be updated frequently enough that
this isn't an issue.