cip-core-image-security.bb:Add libpam-google-authenticator package
Add libpam-google-authenticator package to the security image, in order
achieve the Multi-factor authentication requirement.

This package is already added as a debian dependency in the
security-customizations recipe, and also mentioning it here in the image
recipes to make sure all the packages required for the IEC requirement
is added at one place.

Signed-off-by: Shreyas Karmahe <Shreyas.Karmahe@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>