layer.conf: fix sstate cacheability issue
The LAYERDIR_cip-core contains an absolute path to the layer
at the current location on the fs.
This breaks sstate caching in case the layer is checked out at a
different location on the next build (which is common in CI).

By setting the vardepvalue to a fixed string, this issue is solved.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: whitespace fixup]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>